prueba/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.7
djangorestframework==3.11.0
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
